+++ CMakeLists.txt
@@ -1,4 +1,6 @@
 cmake_minimum_required(VERSION 3.10)
-project(MyProject)
 
-add_executable(MyProject main.cpp)
+project(ffb_example)
+
+add_executable(ffb_example ffb_example.cpp)
+target_link_libraries(ffb_example dinput8 dxguid)
